organize code.  store sorted user list.
